appengine/staging/requirements.txt
Flask==1.1.1
pandas
pylint
google-cloud-storage
google-cloud-bigquery
flasgger